+++ noodles-vcf/src/io/writer/record/info/field/value/array.rs
@@ -110,8 +110,8 @@ mod tests {
         let array = ArrayBuf::Character(vec![Some('n')]);
         t(&mut buf, &array, b"n")?;
 
-        let array = ArrayBuf::Character(vec![Some('n'), Some(':'), None]);
-        t(&mut buf, &array, b"n,%3A,.")?;
+        let array = ArrayBuf::Character(vec![Some('n'), Some(':'), Some('='), None]);
+        t(&mut buf, &array, b"n,:,%3D,.")?;
 
         let array = ArrayBuf::String(vec![Some(String::from("noodles"))]);
         t(&mut buf, &array, b"noodles")?;